Folding evm modules together
