+++ packages/hyperverse-evm/CHANGELOG.md
@@ -1,0 +1,69 @@
+# @decentology/hyperverse-ethereum
+
+## 1.0.9
+
+### Patch Changes
+
+-   Milestone update. Added token and er721 modules
+-   Updated dependencies
+    -   @decentology/unstated-next@1.1.1
+    -   @decentology/web3modal@1.9.6
+    -   @decentology/hyperverse@1.0.9
+
+## 1.0.8
+
+### Patch Changes
+
+-   9df884a: Updated package deps versions for semver
+-   Milestone update. Added support for multi chain swapping
+-   Updated dependencies [9df884a]
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.8
+
+## 1.0.7
+
+### Patch Changes
+
+-   SSL error with SiaSky.net vs FilePortal.org
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.7
+
+## 1.0.6
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+-   38bce19: fixed caching on logout
+-   8183269: Bug fixes for login/logout workflow
+-   Updated dependencies [8183269]
+    -   @decentology/hyperverse@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+-   Handle metamask workflow and stablizing platform
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+-   Version bump after changeset and build
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+-   Milestone change of state management
+-   Updated dependencies
+    -   @decentology/hyperverse@1.0.2
