+++ packages/hyperverse-evm/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@decentology/config/react-library.json",
+  "include": ["source/**/*.ts", "source/**/*.tsx"],
+  "exclude": ["node_modules", "distribution"]
+}
