+++ packages/hyperverse-metis/package.json
@@ -20,6 +20,7 @@
   },
   "dependencies": {
     "@decentology/hyperverse": "^1.0.9",
+    "@decentology/hyperverse-evm": "^1.0.0",
     "@decentology/web3modal": "^1.9.6",
     "@walletconnect/web3-provider": "^1.7.1",
     "ethers": "^5.5.3",
